Enable fuller use of React via not overpacking
NPM has this wonderful feature where it just assumes that anything in the
node_modules directory is supposed to be there and should be preferred over
congruent dependencies specified higher up in the tree. Specifically, this
means that React that was bundled as a devDependency in the module is also
being initalized despite being the same version of React that the app
specifies. So it uses it and gets confused.

So this commit adds a `postbuild` step that copies this project into a sibling
directory -- which is the one the app uses -- and deletes the devDependencies
and peerDependencies so they don't overwrite anything. Sadly, the followon
dependiecies remain, so we'll see what happens there, but for now this works.