Display transaction data in Transaction component
This commit makes the Transaction component display the data presented in the
(as of this writing, fake) transaction response body. It uses the  `date-fns`
module to display the date and the standard `Intl.NumberFormat` to display the
money. The display of the toggle between To Review and Categorized is done
primarily through CSS, since it's all style. The mechanics of it do not affect
its display.