Load Transactions via the API
We need to make sure that we can load Transactions via the API in order for
these components to be useful. We had some issues getting that started, which
is why we were using static data before. But this commit gets the network
access working, so we can load Transactions dynamically.