Add tests to Hello to get them working
Hello is obviously not useful for anything, but it's useful for bootstrapping,
at least. This commit adds a test, but it also adds the intrastructure for
testing, including translating the code via esbuild for jest, jest itself,
react-testing-library, and the mocks necessary to get things rolling. On, and
an example test.