Load and Display Categories for Transactions
In order to categorize transactions, we need to load the list of possible
categories. This list is static for the purposes of categorization. It is
custom to the business and can change, but not during this process. So we load
it once upon login and that should be good for now.

The display is putting the categories (and sub-categories) into a `select` and
split into `option`s and `optgroup`s. This is not super elegant, design-wise,
but will be iterated on for both design and layout.