Filter categorization status
The RadioButtonGroup for chaging which Transactions are displayed already
exists. This commit hooks it up so that Transactions are filtered based on the
choice made in the button. CATEGORIZED, JOURNALING, and SPLIT Transactions are
considered Categorized. READY_FOR_INPUT and LAYER_REVIEW are not. PENDING is
considered such a transient state that it's not worth worrying about.