CategoryMenu creates solo optgroup for lone items
In order to present a more consistent menu, an item that does not have
children but when there is still room in the `maxDepth` should create a
subgroup for itself.

Also, change the way items are default-selected because React greatly prefers
this one, using `defaultValue` instead of selecting one manually.