Layer API function implementations
This commit adds the `src/api/` directory, which contains functions related to
getting data from the Layer API. Aside from getting authorized, the point of
focus is the `get` function in `src/api/layer/authenticated_http.ts`. This
function is curried and works in stages.

It is called with its argument being a function that generates a URL from an
object. This makes the return value suitable to be used as an API function
base, since it can only access one endpoint (with variation for arguments).
The next step is defined when it is given to `useSWR`, where we give it an
`access_token` and any arguments it needs to generate the URL. This defnition
returns the final function where everything is locked in and ready to go. When
`useSWR` calls this function (with arguments, but we really don't care about
them at this point), it will execute the fetch using the values that have been
locked in previously.