Convert LayerProvider to new API functions
The LayerProvider was doing what it needed to, but we have shiny new API
functions now, so we should use them.

Note: This is probably something future developers will know, but: Since we're
defining the Provider here, we don't have a Provider above us in the trew. So
even though we're filling in the state of the Context here, we can't use
`useLayerContext` or `useContet(LayerContext)` here. They won't provide the
values we expect. Categories needs to be called "directly" as defined here if
we want it to be called inside this Provider.