Rename Transaction to BankTransaction everywhere
From @doneel on PR #8: "would prefer we refer to these as bankTransactions
specifically not just transactions. I didn't make that distinction in the api
early on and it's already annoying."

So, this changes that earlier rather than later. Every mention of
`Transaction` in this context (which is all of them) gets changed to a
context-appropriate version of `BankTransaction`.