Implement the BEM naming convention
BEM naming conventions for CSS styling are aimed at being an easy and
consistent way to name CSS styles for an entire project. Here, we were already
feeling the pinch of having some confusing and arbitrarily named classes and
styles, so we figured how would be a good time to use a consistent naming
convention. BEM comes highly recommended by other members of the team.

https://getbem.com/