Let useBankTransactions update data in cache
After we update the Transaction via the API call, we should update the local
cache, so we don't have to fetch all the transactions again. This commit adds
a `mutateOne` function to the hook's return value so we can do just that. It
accepts a (partial) BankTransaction and updates that entry in the cache.