Get the save buttons working to save Categories
This commit makes the Save button on the ExpandedBankTransactionRow and the
BankTransactionRow work so that the BankTransactions can be categorized. It
does this by moving the API call to catgorize the transaction into the
useBankTransactions hook instead of keeping the two (API + cache mutation)
separate.

This commit also adds parameterized types to the `get` and `put` functions so
they can be more explicit when turned into proper API calls.

As of this commit, the uswSWR hook fires an extra time, which leads to some
apparent data duplication. The cause of this has not yet been determined.