Improve the data returned from useBankTransactions
useBankTransactions returns the raw data from the API endpoint. This includes
both the `data` (BankTransactions) and the `meta` (metadata, which is
pagination info (as of this commit)). We change this to return these two
pieces of data separately, named, respectively, `data` and `metadata`. Also,
adjust the types to match and the test to reflect what's actually happening.