Add Profit and Loss report API, types, and hook
In order to display the Profit and Loss report, we need to be able to get it
from the server as well as determine what is inside it. The `ProfitAndLoss`
type was converted from the Kotlin server code. The API endpoint uses the
standard `get` function to create its endpoint and puts the required
`startDate` and `endDate` in the URL so they can't be forgotten.

This also adds the UseProfitAndLoss hook, which brings the Profit and Loss
data into easy reach. By default, it loads the data for the current month, but
can be modified via internal functions. Props would normally be preferable,
but the way this is going to be used is by having a container component and
sub-components, so they won't be able to change the container via props
anyway.