Implement Profit and Loss components
This commit actually implements the comonents that display the P&L data. The
markup itself is a simple two-column table. The non-calculated values in the
P&L report have line items for expanding and explaining where the data came
from, and they can be displayed in a recursive fashion. The calculated fields
are shown by making them look like the other fields.