Implement Date Picker and pull apart ProfitAndLoss
The design of the ProfitAndLoss component should be modular, allowing the
client the ability to implement as much or as little as they want and allowing
somewhat of their own design. To this end, the implementation of the Date
Picker presents an opportunity to separate parts of the full ProfitAndLoss to
be split apart and recomposed.

The uppermost ProfitAndLoss component creates a Context for the
`useProfitAndLoss`'s data to live in while also giving the sub-components a
logical home in the HTML. This allows the sub-components to share the same P&L
data (and date range) while also netting the benefits of being under the
useSWR hook.