ProfitAndLoss Chart using recharts
The top widget of the ProfitAndLoss design is a chart that puts income and
expewnses against time, grouped by month. This commit presents this chart
using recharts.

As of this commit, there is no endpoint on the Layer side that presents this
data directly, so we have to request the previous 12 P&L reports from the
server and synthesize the report data from them.