Create useWindowSize to trigger size repaints
The ProfitAndLoss Chart only takes absolute number sizes for its width, so if
we want to be able to fit the chart to the width of the ProfitAndLoss
component (or allow for percent/relative sizes in the future), we need to be
able to measure its containing element. But since the size of its containing
element may change when the window is resized, we can use window-size-events
to be able to trigger our `useEffect`s or `useLayoutEffect`s. This is a handy
way of wrapping that in a hook. The original code was found on StackOverflow
and since it wasn't worth rewriting, it was copied and a credit link was put
into a comment.