Properly respond to TypeScript errors
In which we reveal a terrible secret: The TypeScript types were not being
generated properly!

The `noEmit` option was turned on in the tsconfig, since we don't want running
`tsc` to generate a bunch of `js` files. This was useful for errant file
creation, but was less ideal in that `npm-dts` does not generate type files
when this option is `true`. `npm-dts` also does not generate errors when being
run as part of the build command. Running it from the command line surfaces
those errors.

And there were a bunch of errors! The rest of this commit is in place to
address them.