Add a title, Legend, and move the indictor below
This gives the whole P&L a title, it moves the Legend to the top and styles
it, and it moves the indicator to the bottom of the chart where it's more
useful. It also updates some styles that were updated for the Table, including
making the section titles bold, adding thicker borders on calculated rows,
and making sure numbers had commas for thousands places.

Generally, this now looks much more like the current state of the Figma designs.