BalanceSheet API endpoint and hooks
In order to use the Balance Sheet API, we need to define both the endpoint and
a hook to make it straightforward to access. These are fairly standard and
don't need much explanation by now.

NOTE: The data is static at the moment because of the limittaions of the
actual API. This static data is more in line with the eventual output, so it's
being used for now.