Show the contents of the Balance Sheet
The Balance Sheet looks very similar to the Profit and Loss Table. So close,
in fact, that most of this code was just straight lifted from it, but with
style tweaks. Anyway, the structure is basically the same: The main container
has the title and some controls, while the display itself is showing a
heirarchical view of nested line items. One change is that this one has two
container rows -- Assets and Liabilities & Equity -- which do not display
summary totals.