Implement the Balance Sheet Date Picker Button
The date picker we're using is just the standard `input type="date"` picker as
supplied by the browser. But since the input element itself has variations
between browsers and because we want it to appear like a button, we're doing a
bit of trckery to get it to look right. We're keeping the element visible in
case screen readers care (and also beause we want the picker portion of it),
but we're hiding it behind an actual button element. This will remove
variation while keeping usability and accessibility high (as both elements are
still visible and usable).

Also, despite the availability of `valueAsDate` on the element itself, as
presented by the HTML standard, the date appears as a local Date object once
parsed as UTC, which means if we select 1/1/2024, we recieve an object that is
12/31/2023T19:00-0500, which does not work for our purposes. Parsing the date
with `date-fns` returns a date that is 1/1/2024T00:00-0500. The difference is
one parsed in UTC and the translated to local time versus one parsed already
in local time. While it's arguable which one is "correct", the latter is the
one that is usable for our purposes without futher modification.