Parameterize auth configuration
This commit moves the `appId`, etc from being defined as constants in
`src/api/layer/authenticate.ts` and moves it to being defined as a prop on the
LayerProvider. This is a step toward making the module publishable, since
having static auth config wouldn't be terribly useful to people.

Also, it shifts around some of the code involving mocking `fetch` since that
wasn't working for some reason.