Animate the chart Indicator when selection changes
Recharts allows us to specify content for the Bars in a chart using a
LabelList component. This component can customize its content to show what
we would like to show. In this case, we draw a box below the X axis only on
the selected month so we can indicate selection -- hence "Indicator".

However, this component does not exist beween selections. In goes away when a
new month is clicked and a new one is created. This hampers attempts to
animate it. Thw workaround is to maintain the posiiton of the Indicator so tat
when a new month is selected, we create the Indicator component using the old
X coordinate. `useEffect` always fires after the window has painted -- which
means the X coordinate is "officially" on the screen. We then update the saved
X coordinate to the one it's supposed to be, according to the selected month.
That causes another render, which updates the X coordinate to that correct
one. Finally the browser sees the update and CSS-animates it into place.