Keep displaying markup even if data isn't there
The ProfitAndLossSummaries should stick around even if they don't have data.
Seeing them blip in an out of existence is jarring. If there is no data for
them to display, they should display `$-.--` so their structre sticks around.

This also applies to the ProfitAndLossTable. We need a "blank" document that
replicates the things we need to show, but with `NaN` as values. We can use
`NaN` to denote that dollar formatting should be `$-.--` instead of `$0.00`.