Animate hiding/revealing Balance Sheet row
When the Balance Sheet row is clicked, it toggles its visibility. This commit
animates that change in the form of rotating the Chevron displayed next to the
display_name.

Kt also animates a show/hide of the contents of the child rows. It does this
by animating the `grid-template-rows` CSS property of a subgrid of a subgrid.
This lets the animation happen on a "single celled grid" that is actually a
subgrid of the larger table.