Show summary total of each Balance Sheet group
Each line item in the Balance Sheet comes with a total for it's consitituent
line items. When the line is expanded, we should show the total below the lkst
of lines in a "Total of <name>" line. But when it's collapsed, it should show
in the normal value spot for its line, on the right side of the table.

Also, the Assets and the Liabilities & Equity sections do not (currently) have
their own summations, so we should not show a line for them yet.