Extract all styles into their own logical files
This project started with having all SCSS in the `index.scss` file. This was
(mostly) find for the single component that was in play at the time, but since
we have three major components (and a handful of cross-cutting minor ones), we
should make sure they are properly broken out into files that apply to only
them, and leave the index for being an index.