Animate hiding BankTransactionRow post-categorize
The actual mechanics of the transitions were set up in the previous commit.
This commit implements the triggering of those transitions. When a
BankTransaction is categorized, we update the SWR cache and tell it not to
revalidate (this is important or it would refresh the list from the server).
As we update the cache, we also set `recently_categorized` to true on the
transaction. The BankTransactions list now includes `recently_categorized`
transactions in the "In Review" list for display. This lets us add a new class
to those transactions that will slide them up like the BalanceSheetRows and
ExpandedBankTransactionRows.

One quirk is that we add `top` to the BankTransactionRow collapsing transition
and `bottom` to the ExpandedBankTransactionRow opening transition. This is
because we need to know which transition is ending in the `onTransitionEnd`
handler, and each property in the transition sends an event. They don't do
anything on their own.