Implement (ugly) New Account Form
This component allows the user to create an Account as a child of another
Account. It doesn't do it beautifully, but it works. It obtains a list of
Accounts that can be parents by getting the whole Chart of Accounts and
flattening it.

We also follow the pattern of putting the modifying functions in the hook,
where it can access internals, and passing them out to the callers so
modifications can be made without needing details manually.