Expand HTTP request repertoire to add POST
Since we're generating functions that lock in details about HTTP requests, it
seems fitting to do more of that! This commit makes a generator function,
called `request`, that generates more specific functions that deal with one
HTTP verb. From the caller's perspective, it's the same, but it simply allows
us to state that the content of a `put` and a `post` function aren't really
all that dissimilar (varying only in the verb).

It also changes the order of yhe parameterized types of the non-GET functions,
swapping it around to `<Return, Body, Params>`, which always leaves `Return`
first but also puts `Params` last as they're least likely to need to be
specified.