Adds createAccount endpoint
This adds the endpoint for creating an account. The only interesting thing to
say about this commit is that the return value of the Chart of Accounts
endpoint returns keys in `camelCase` while the return value of Create Account
returns keys in `snake_case`, and so we can't make a 1:1 drop in replacement
in the SWR cache.