Do not reauthenticte if the token has not expired
The aiuthentication request returns an `expires_in` value. That means the
token is good for that long. We should not retrieve new tokens until that span
of time is over.

This commit calculates the time when the token will expire and stores it in
the OAuthResponse to make an ExpiringOAuthResponse. It then compares that time
to the current time in order to determine if it should request authentication
again.